// src/api/types/commit.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RefChip } from "./ref_chip";

export type Commit = { hash: string, short_hash: string, parents: Array<string>, author_name: string, author_email: string, author_date: bigint, committer_name: string, committer_email: string, committer_date: bigint, refs: Array<RefChip>, subject: string, body: string, };